Refactor using more custom types
Abstract more of the getTag parsing functions. This should allow for
some more interisting compositions.

for now using tagFieldToMaybeString in the json object but this should
simple to map.

Rename valueToStringMay to valueToString as Maybe values are assummed
at this point.

functionality should still be the same as the previous commit.