Add support for multi-value tags & more refactor
- the import and export lists are not final, but just explicitly add
everything for now

- rename cs to currentSong and nextPlaylistSong to nextSong, "current"
and "next" are assumed to be in the current playlist because that's
how MPD handles 'tracklists'.

- Create Network.MPD.JSON module for json-related functions like
building key-value pairs. Thanks to a magic in toJSON function it is
able to convert the custom TagField type to supported `Value's for
`Data.Aeson.object'.

- Remove deprecated functions and use expanded logic of the TagField
type to handle both `Maybe String' and `Maybe [String]'.

Removed headMay function in favour of builtin (!?) which is the safe
version of (!!)

getTagNextSong was merged into getTag by using an expanded type logic
in TagField.

processSong was unnecessary.

- BIG!