fix reverse order of multi-value tags
for some reason multi-value tags are presented in reverse order by
libmpd. For example if MPD reads a `artists' tag with value "John;
Mike; Hannah", libmpd would return a list with value
"["Hannah", "Mike", "John"]". A simple `reverse' on the parsed list
fixes it since it's just actually reversed and not randomized.